refactor(demo): remove src folder for remotion app
