+++ apps/demo/app/layout.tsx
@@ -1,4 +1,4 @@
-import "../../styles/global.css";
+import "../styles/global.css";
 import type { Metadata, Viewport } from "next";
 
 export const metadata: Metadata = {
