+++ apps/demo/remotion/components/visual/SignUpFormVisual.tsx
