+++ apps/demo/remotion/index.ts
@@ -1,4 +1,4 @@
-import "../../styles/global.css";
+import "../styles/global.css";
 
 import { registerRoot } from "remotion";
